[Move] Post Likes Count 함수 - Formatter 파일로 이동 #15
